agent/pyproject.toml
[project]
name = "community-assistant"
version = "0.1.0"
description = "Multi-agent Discord assistant for the Code with Ahsan community, powered by Google ADK."
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["community_assistant"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --tb=short"
